Adding a check before installing Helm in the install_helm function of the helm utility script.

+++ scripts/utilities/helm.sh
@@ -4,14 +4,16 @@
 #
 
 function install_helm () {
-  info "Installing Helm ..."
+  if ! command -v "${__zimagi_binary_dir}/helm" > /dev/null; then
+    info "Installing Helm ..."
 
-  info "Downloading Helm installer ..."
-  curl -o /tmp/helm_install.sh https://raw.githubusercontent.com/helm/helm/master/scripts/get 2>/dev/null
-  chmod 700 /tmp/helm_install.sh
+    info "Downloading Helm installer ..."
+    curl -o /tmp/helm_install.sh https://raw.githubusercontent.com/helm/helm/master/scripts/get 2>/dev/null
+    chmod 700 /tmp/helm_install.sh
 
-  info "Running Helm installer ..."
-  export HELM_INSTALL_DIR="${__zimagi_binary_dir}"
-  /tmp/helm_install.sh -v "v${HELM_VERSION}" --no-sudo >/dev/null 2>&1
-  rm -f /tmp/helm_install.sh
+    info "Running Helm installer ..."
+    export HELM_INSTALL_DIR="${__zimagi_binary_dir}"
+    /tmp/helm_install.sh -v "v${HELM_VERSION}" --no-sudo >/dev/null 2>&1
+    rm -f /tmp/helm_install.sh
+  fi
 }
